create canary module
